+++ src/server/WacAllowHttpHandler.ts
@@ -71,6 +71,8 @@ export class WacAllowHttpHandler extends OperationHttpHandler {
       } else {
         // Need to determine public permissions
         this.logger.debug('Determining public permissions');
+        // Note that this call can potentially create a new lock on a resource that is already locked,
+        // so a locker that allows multiple read locks on the same resource is required.
         const permissionMap = await this.permissionReader.handleSafe({ credentials: {}, requestedModes });
         everyone = permissionMap.get(operation.target) ?? {};
       }